feat: implement campaign detail page components and layout structure
