+++ src/app/(marketing)/about/page.tsx
@@ -1,5 +1,5 @@
 
-'use client';
+
 
 import { CheckCircle, Gift, PlusCircle, Rocket, Users, Zap, ShieldCheck, ShieldOff, Bot } from 'lucide-react';
 import { Card, CardContent, CardHeader, CardTitle, CardDescription } from '@/components/ui/card';
